Move call function to Provider (#63)
* Move call function to provider

* Fix compilation

* Fix clippy
