+++ src/zks_wallet/wallet.rs
@@ -6,13 +6,12 @@ use crate::{
     eip712::{hash_bytecode, Eip712Meta, Eip712TransactionRequest},
     zks_provider::ZKSProvider,
     zks_utils::{
-        self, is_precompile, CONTRACT_DEPLOYER_ADDR, DEPOSIT_GAS_PER_PUBDATA_LIMIT, EIP712_TX_TYPE,
-        ERA_CHAIN_ID, ETH_CHAIN_ID, RECOMMENDED_DEPOSIT_L1_GAS_LIMIT,
-        RECOMMENDED_DEPOSIT_L2_GAS_LIMIT,
+        self, CONTRACT_DEPLOYER_ADDR, DEPOSIT_GAS_PER_PUBDATA_LIMIT, EIP712_TX_TYPE, ERA_CHAIN_ID,
+        ETH_CHAIN_ID, RECOMMENDED_DEPOSIT_L1_GAS_LIMIT, RECOMMENDED_DEPOSIT_L2_GAS_LIMIT,
     },
 };
 use ethers::{
-    abi::{decode, encode, Abi, HumanReadableParser, ParamType, Token, Tokenizable, Tokenize},
+    abi::{decode, Abi, ParamType, Token, Tokenizable, Tokenize},
     prelude::{
         encode_function_data,
         k256::{
@@ -398,7 +397,7 @@ where
             if let Some(contract_dependencies) = contract_dependencies {
                 factory_deps.extend(contract_dependencies);
             }
-            factory_deps.push(contract_bytecode.clone().to_vec());
+            factory_deps.push(contract_bytecode.to_vec());
             factory_deps
         });
 
@@ -432,7 +431,7 @@ where
                 })?;
                 // TODO: User could provide this instead of defaulting.
                 let salt = [0_u8; 32];
-                let bytecode_hash = hash_bytecode(&contract_bytecode)?;
+                let bytecode_hash = hash_bytecode(contract_bytecode)?;
                 let call_data = Bytes::default();
 
                 encode_function_data(create, (salt, bytecode_hash, call_data))?
@@ -572,67 +571,6 @@ where
             .ok_or(ZKSWalletError::CustomError(
                 "no transaction receipt".to_owned(),
             ))
-    }
-
-    pub async fn call(
-        &self,
-        contract_address: Address,
-        function_signature: &str,
-        function_parameters: Option<Vec<String>>,
-    ) -> Result<Vec<Token>, ZKSWalletError<M, D>>
-    where
-        M: ZKSProvider,
-    {
-        let era_provider = match &self.era_provider {
-            Some(era_provider) => era_provider,
-            None => return Err(ZKSWalletError::CustomError("no era provider".to_owned())),
-        };
-
-        // Note: We couldn't implement ZKSWalletError::LexerError because ethers-rs's LexerError is not exposed.
-        let function = if contract_address == zks_utils::ECADD_PRECOMPILE_ADDRESS {
-            zks_utils::ec_add_function()
-        } else {
-            HumanReadableParser::parse_function(function_signature)
-                .map_err(|e| ZKSWalletError::CustomError(e.to_string()))?
-        };
-        let function_args = if let Some(function_args) = function_parameters {
-            function
-                .decode_input(&*zks_utils::encode_args(&function, &function_args)?)
-                .map_err(|e| ZKSWalletError::CustomError(e.to_string()))?
-        } else {
-            vec![]
-        };
-
-        log::info!("{function_args:?}");
-
-        let request: Eip1559TransactionRequest =
-            Eip1559TransactionRequest::new().to(contract_address).data(
-                match (!function_args.is_empty(), is_precompile(contract_address)) {
-                    // The contract to call is a precompile with arguments.
-                    (true, true) => encode(&function_args),
-                    // The contract to call is a regular contract with arguments.
-                    (true, false) => function
-                        .encode_input(&function_args)
-                        .map_err(|e| ZKSWalletError::CustomError(e.to_string()))?,
-                    // The contract to call is a precompile without arguments.
-                    (false, true) => Default::default(),
-                    // The contract to call is a regular contract without arguments.
-                    (false, false) => function.short_signature().into(),
-                },
-            );
-
-        let transaction: TypedTransaction = request.into();
-
-        let encoded_output = era_provider.call(&transaction, None).await?;
-        let decoded_output = function.decode_output(&encoded_output).map_err(|e| {
-            ZKSWalletError::CustomError(format!("failed to decode output: {e}\n{encoded_output}"))
-        })?;
-
-        Ok(if decoded_output.is_empty() {
-            encoded_output.into_tokens()
-        } else {
-            decoded_output
-        })
     }
 
     pub async fn withdraw(
@@ -814,13 +752,12 @@ mod zks_signer_tests {
     use crate::compile::project::ZKProject;
     use crate::test_utils::*;
     use crate::zks_wallet::ZKSWallet;
-    use ethers::abi::{Token, Tokenize};
     use ethers::providers::Middleware;
     use ethers::signers::LocalWallet;
     use ethers::solc::info::ContractInfo;
     use ethers::solc::{Project, ProjectPathsConfig};
+    use ethers::types::Address;
     use ethers::types::U256;
-    use ethers::types::{Address, Bytes};
     use ethers::utils::parse_units;
     use std::str::FromStr;
 
@@ -1029,77 +966,6 @@ mod zks_signer_tests {
         let recovered_bytecode = era_provider.get_code(contract_address, None).await.unwrap();
 
         assert_eq!(compiled_bytecode, recovered_bytecode);
-    }
-
-    #[tokio::test]
-    async fn test_call_view_function_with_no_parameters() {
-        // Deploying a test contract
-        let deployer_private_key =
-            "7726827caac94a7f9e1b160f7ea819f172f7b6f9d2a97f992c38edeab82d4110";
-        let era_provider = era_provider();
-        let wallet = LocalWallet::from_str(deployer_private_key).unwrap();
-        let zk_wallet = ZKSWallet::new(wallet, None, Some(era_provider.clone()), None).unwrap();
-
-        let contract_address = zk_wallet
-            .deploy::<Token>("src/compile/test_contracts/test/src/Test.sol", "Test", None)
-            .await
-            .unwrap();
-
-        // Making the call to the contract function
-        let deployer_private_key =
-            "0x28a574ab2de8a00364d5dd4b07c4f2f574ef7fcc2a86a197f65abaec836d1959";
-        let wallet = LocalWallet::from_str(deployer_private_key).unwrap();
-        let zk_wallet = ZKSWallet::new(wallet, None, Some(era_provider.clone()), None).unwrap();
-
-        let output = zk_wallet
-            .call(contract_address, "str_out()(string)", None)
-            .await
-            .unwrap();
-
-        assert_eq!(output, String::from("Hello World!").into_tokens());
-    }
-
-    #[tokio::test]
-    async fn test_call_view_function_with_arguments() {
-        // Deploying a test contract
-        let deployer_private_key =
-            "7726827caac94a7f9e1b160f7ea819f172f7b6f9d2a97f992c38edeab82d4110";
-        let era_provider = era_provider();
-        let wallet = LocalWallet::from_str(deployer_private_key).unwrap();
-        let zk_wallet = ZKSWallet::new(wallet, None, Some(era_provider.clone()), None).unwrap();
-
-        let contract_address = zk_wallet
-            .deploy::<Token>("src/compile/test_contracts/test/src/Test.sol", "Test", None)
-            .await
-            .unwrap();
-
-        let no_return_type_output = zk_wallet
-            .call(
-                contract_address,
-                "plus_one(uint256)",
-                Some(vec!["1".to_owned()]),
-            )
-            .await
-            .unwrap();
-
-        let known_return_type_output = zk_wallet
-            .call(
-                contract_address,
-                "plus_one(uint256)(uint256)",
-                Some(vec!["1".to_owned()]),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(
-            no_return_type_output,
-            Bytes::from([
-                0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-                0, 0, 0, 2
-            ])
-            .into_tokens()
-        );
-        assert_eq!(known_return_type_output, U256::from(2_u64).into_tokens());
     }
 
     #[tokio::test]
